+++ x/testutil/cli/query.go
@@ -11,10 +11,10 @@ import (
 	tmcli "github.com/tendermint/tendermint/libs/cli"
 
 	"github.com/NibiruChain/nibiru/x/common"
+	oraclecli "github.com/NibiruChain/nibiru/x/oracle/client/cli"
+	oracletypes "github.com/NibiruChain/nibiru/x/oracle/types"
 	perpcli "github.com/NibiruChain/nibiru/x/perp/client/cli"
 	perptypes "github.com/NibiruChain/nibiru/x/perp/types"
-	pricefeedcli "github.com/NibiruChain/nibiru/x/pricefeed/client/cli"
-	pricefeedtypes "github.com/NibiruChain/nibiru/x/pricefeed/types"
 	vpoolcli "github.com/NibiruChain/nibiru/x/vpool/client/cli"
 	vpooltypes "github.com/NibiruChain/nibiru/x/vpool/types"
 )
@@ -83,6 +83,14 @@ func QueryVpoolReserveAssets(clientCtx client.Context, pair common.AssetPair,
 	return &queryResp, nil
 }
 
+func QueryOracleExchangeRate(clientCtx client.Context, pair string) (*oracletypes.QueryExchangeRateResponse, error) {
+	var queryResp oracletypes.QueryExchangeRateResponse
+	if err := ExecQuery(clientCtx, oraclecli.GetCmdQueryExchangeRates(), []string{pair}, &queryResp); err != nil {
+		return nil, err
+	}
+	return &queryResp, nil
+}
+
 func QueryBaseAssetPrice(clientCtx client.Context, pair common.AssetPair, direction string, amount string) (*vpooltypes.QueryBaseAssetPriceResponse, error) {
 	var queryResp vpooltypes.QueryBaseAssetPriceResponse
 	if err := ExecQuery(clientCtx, vpoolcli.CmdGetBaseAssetPrice(), []string{pair.String(), direction, amount}, &queryResp); err != nil {
@@ -104,22 +112,5 @@ func QueryCumulativePremiumFraction(clientCtx client.Context, pair common.AssetP
 	if err := ExecQuery(clientCtx, perpcli.CmdQueryCumulativePremiumFraction(), []string{pair.String()}, &queryResp); err != nil {
 		return nil, err
 	}
-	return &queryResp, nil
-}
-
-func QueryPrice(clientCtx client.Context, pairID string) (*pricefeedtypes.QueryPriceResponse, error) {
-	var queryResp pricefeedtypes.QueryPriceResponse
-	if err := ExecQuery(clientCtx, pricefeedcli.CmdQueryPrice(), []string{pairID}, &queryResp); err != nil {
-		return nil, err
-	}
-	return &queryResp, nil
-}
-
-func QueryRawPrice(clientCtx client.Context, pairID string) (*pricefeedtypes.QueryRawPricesResponse, error) {
-	var queryResp pricefeedtypes.QueryRawPricesResponse
-	if err := ExecQuery(clientCtx, pricefeedcli.CmdQueryRawPrices(), []string{pairID}, &queryResp); err != nil {
-		return nil, err
-	}
-
 	return &queryResp, nil
 }